js: fixed alert color being overridden
The alert notifying the user that the bootstrap has failed was
overriding the type of any previous alert that were on the page.
This was causing some confusing to some users.

Now it creates an alert everytime we check if there's a failed
bootstrapped node without interfering other alerts.

Also removed unnecessary stuff from the notification partial.

Fixes bsc#1069258